requirements.txt
pandas==2.0.3
pandasql==0.7.3
jupyter==1.0.0
psycopg2-binary==2.9.9
pyarrow==17.0.0
fastparquet==2024.5.0
datetime
apache-airflow==2.10.1
apache-airflow-providers-postgres==5.1.0